Make dependency on babel optional
As far as Python packages go, babel is quite large weighing in at 29M.
This is due to it containing a very large amount of locale data.

As babel is only used by the PhonePrefixSelect, which isn't used by all
library users, make the dependency optional.

For library users that vendor or bundle django-phonenumber-field and its
dependencies, this greatly cuts down on the final size.